+++ electron-builder.config.js
@@ -139,6 +139,7 @@ module.exports = {
     'node_modules/@img/**/*',
     'node_modules/tesseract.js/**/*',
     'node_modules/tesseract.js-core/**/*',
-    'node_modules/detect-libc/**/*'
+    'node_modules/detect-libc/**/*',
+    'node_modules/semver/**/*'
   ]
 }